aider: Fixed error in user_connected function by replacing then with chain to append Once stream to existing stream.
# Aider chat conversation:

USER: I ran this command:

 cargo build

And got this output:

   Compiling agenthost v0.1.0 (/home/runvnc/agenthost)
error[E0277]: `futures::stream::Once<[async block@src/api/mod.rs:254:31: 254:107]>` is not a future
   --> src/api/mod.rs:252:8
    |
252 |     }).then(move |_| {
    |        ^^^^ `futures::stream::Once<[async block@src/api/mod.rs:254:31: 254:107]>` is not a future
    |
    = help: the trait `futures::Future` is not implemented for `futures::stream::Once<[async block@src/api/mod.rs:254:31: 254:107]>`
    = note: futures::stream::Once<[async block@src/api/mod.rs:254:31: 254:107]> must be a future or must implement `IntoFuture` to be awaited
note: required by a bound in `futures::StreamExt::then`
   --> /home/runvnc/.cargo/registry/src/index.crates.io-6f17d22bba15001f/futures-util-0.3.28/src/stream/stream/mod.rs:485:14
    |
482 |     fn then<Fut, F>(self, f: F) -> Then<Self, Fut, F>
    |        ---- required by a bound in this associated function
...
485 |         Fut: Future,
    |              ^^^^^^ required by this bound in `StreamExt::then`

error[E0277]: `futures::stream::Once<[async block@src/api/mod.rs:254:31: 254:107]>` is not a future
   --> src/api/mod.rs:203:6
    |
203 | ) -> impl Stream<Item = Result<Event, Infallible>> + Send + 'static {
    |      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ `futures::stream::Once<[async block@src/api/mod.rs:254:31: 254:107]>` is not a future
    |
    = help: the trait `futures::Future` is not implemented for `futures::stream::Once<[async block@src/api/mod.rs:254:31: 254:107]>`
    = note: futures::stream::Once<[async block@src/api/mod.rs:254:31: 254:107]> must be a future or must implement `IntoFuture` to be awaited
    = help: the trait `Stream` is implemented for `futures::stream::Then<St, Fut, F>`
    = note: required for `Then<Map<UnboundedReceiverStream<ChatUIMessage>, [closure@mod.rs:219:25]>, Once<...>, ...>` to implement `Stream`
    = note: the full type name has been written to '/home/runvnc/agenthost/target/debug/deps/agenthost-bfe4d02096c65793.long-type-12729391580998793675.txt'

For more information about this error, try `rustc --explain E0277`.
error: could not compile `agenthost` (bin "agenthost") due to 2 previous errors

ASSISTANT: The error message indicates that the `then` combinator is expecting a `Future`, but it is being given a `Stream`. The `futures::stream::Once` is a stream, not a future, which is why the error is occurring. The `then` combinator is used to chain a future computation onto the results of a stream, but it seems like the intention here is to append a single event to the stream when the client disconnects.

To fix this, you should use the `chain` method to append the `Once` stream to the existing stream, rather than using `then`. Here's how you can modify the `user_connected` function to fix the issue:

src/api/mod.rs
```rust
<<<<<<< SEARCH
}).then(move |_| {
    println!(">>>>>>>>>>>>>>>>>>>>> Client disconnected");
    futures::stream::once(async move { Ok(Event::default().event("disconnect").data("disconnected")) })
});
=======
}).chain(futures::stream::once(async move { Ok(Event::default().event("disconnect").data("disconnected")) }))
>>>>>>> REPLACE
```

In this *SEARCH/REPLACE* block, I replaced the incorrect usage of `then` with `chain` to append the `Once` stream to the existing stream.

Please apply this change to your `src/api/mod.rs` file and then try to compile your project again with `cargo build`.