aider: Changed iteration over messages to iterate over references and removed unnecessary cloning of messages in the `save()` function.
# Aider chat conversation:

ASSISTANT: Yes, you can avoid cloning by using an iterator to iterate over the references to the messages in the vector. This way, you're not creating a new copy of the messages, but just referencing them.

Here's how you can modify the `save()` function:

src/chatlog/mod.rs
```rust
<<<<<<< SEARCH
for msg in self.messages {
    if i > 0 { outjson.push_str(",\n") }
    let json = serialize_message(&msg.message.clone());
    outjson.push_str(&json);
    i += 1;
}
=======
for msg in &self.messages {
    if i > 0 { outjson.push_str(",\n") }
    let json = serialize_message(&msg.message);
    outjson.push_str(&json);
    i += 1;
}
>>>>>>> REPLACE
```

In the `REPLACE` block, I've changed `for msg in self.messages` to `for msg in &self.messages` to iterate over references to the messages. I've also removed the `.clone()` call from `serialize_message(&msg.message.clone())` because `msg.message` is now a reference and can be passed directly to `serialize_message()`.